Refactored routes, requests now don't need to pass the user id anymore

+++ src/main/kotlin/com/shobhit63/routes/CommentRoutes.kt
@@ -18,47 +18,43 @@ import io.ktor.server.routing.*
 
 fun Route.createComment(
     commentService: CommentService,
-    userService: UserService
-){
+) {
     authenticate {
-        post("/api/comment/create"){
+        post("/api/comment/create") {
             val request = call.receiveOrNull<CreateCommentRequest>() ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ) {
-              when(commentService.createComment(request)){
-                  is CommentService.ValidationEvents.ErrorFieldEmpty -> {
-                      call.respond(
-                          HttpStatusCode.OK,
-                          BasicApiResponse(
-                              successful = false,
-                              message = ApiResponseMessages.FIELDS_BLANK
-                          )
-                      )
-                  }
-                  is CommentService.ValidationEvents.ErrorCommentTooLong -> {
-                      call.respond(
-                          HttpStatusCode.OK,
-                          BasicApiResponse(
-                              successful = false,
-                              message = ApiResponseMessages.COMMENT_TOO_LONG
-                          )
-                      )
-                  }
-                  is CommentService.ValidationEvents.Success -> {
-                      call.respond(
-                          HttpStatusCode.OK,
-                          BasicApiResponse(
-                              successful = true
-                          )
-                      )
-                  }
-              }
+            when (commentService.createComment(request,call.userId)) {
+                is CommentService.ValidationEvents.ErrorFieldEmpty -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.FIELDS_BLANK
+                        )
+                    )
+                }
+
+                is CommentService.ValidationEvents.ErrorCommentTooLong -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = false,
+                            message = ApiResponseMessages.COMMENT_TOO_LONG
+                        )
+                    )
+                }
+
+                is CommentService.ValidationEvents.Success -> {
+                    call.respond(
+                        HttpStatusCode.OK,
+                        BasicApiResponse(
+                            successful = true
+                        )
+                    )
+                }
             }
         }
     }
@@ -75,14 +71,13 @@ fun Route.getCommentsForPost(
                 return@get
             }
             val comments = commentService.getCommentsForPost(postId)
-            call.respond(HttpStatusCode.OK,comments)
+            call.respond(HttpStatusCode.OK, comments)
         }
     }
 }
 
 fun Route.deleteComment(
     commentService: CommentService,
-    userService: UserService,
     likeService: LikeService
 ) {
     authenticate {
@@ -92,19 +87,20 @@ fun Route.deleteComment(
                 return@delete
             }
 
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ){
-                val deleted = commentService.deleteComment(request.commentId)
-                if(deleted) {
-                    likeService.deleteLikesForParent(request.commentId)
-                    call.respond(HttpStatusCode.OK,BasicApiResponse(successful = true))
-                }else{
-                    call.respond(HttpStatusCode.NotFound,BasicApiResponse(successful = false))
-
-                }
+            val comment = commentService.getCommentById(request.commentId)
+            if(comment?.userId != call.userId) {
+                call.respond(HttpStatusCode.Unauthorized)
+                return@delete
             }
+            val deleted = commentService.deleteComment(request.commentId)
+            if (deleted) {
+                likeService.deleteLikesForParent(request.commentId)
+                call.respond(HttpStatusCode.OK, BasicApiResponse(successful = true))
+            } else {
+                call.respond(HttpStatusCode.NotFound, BasicApiResponse(successful = false))
+
+            }
+
         }
     }
 }
