+++ src/main/kotlin/com/shobhit63/routes/PostRoutes.kt
@@ -10,7 +10,6 @@ import com.shobhit63.util.ApiResponseMessages
 import com.shobhit63.util.Constants.DEFAULT_POST_PAGE_SIZE
 import com.shobhit63.util.QueryParams.PARAM_PAGE
 import com.shobhit63.util.QueryParams.PARAM_PAGE_SIZE
-import com.shobhit63.util.QueryParams.PARAM_USER_ID
 import io.ktor.http.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
@@ -20,7 +19,6 @@ import io.ktor.server.routing.*
 
 fun Route.createPost(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         post("/api/post/create") {
@@ -28,29 +26,23 @@ fun Route.createPost(
                 call.respond(HttpStatusCode.BadRequest)
                 return@post
             }
-
-            ifEmailBelongsToUser(
-                userId = request.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ) {
-                val didUserExist = postService.createPostIfUserExists(request)
-                if (!didUserExist) {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = false,
-                            message = ApiResponseMessages.USER_NOT_FOUND
-                        )
+            val userId = call.userId
+            val didUserExist = postService.createPostIfUserExists(request, userId)
+            if (!didUserExist) {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = false,
+                        message = ApiResponseMessages.USER_NOT_FOUND
                     )
-                } else {
-                    call.respond(
-                        HttpStatusCode.OK,
-                        BasicApiResponse(
-                            successful = true,
-                        )
+                )
+            } else {
+                call.respond(
+                    HttpStatusCode.OK,
+                    BasicApiResponse(
+                        successful = true,
                     )
-                }
-
+                )
             }
 
 
@@ -61,27 +53,17 @@ fun Route.createPost(
 
 fun Route.getPostsForFollows(
     postService: PostService,
-    userService: UserService
 ) {
     authenticate {
         get("api/post/get") {
-            val userId = call.parameters[PARAM_USER_ID] ?: kotlin.run {
-                call.respond(HttpStatusCode.BadRequest)
-                return@get
-            }
             val page = call.parameters[PARAM_PAGE]?.toIntOrNull() ?: 0
             val pageSize = call.parameters[PARAM_PAGE_SIZE]?.toIntOrNull() ?: DEFAULT_POST_PAGE_SIZE
 
-            ifEmailBelongsToUser(
-                userId = userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ) {
-                val posts = postService.getPostsForFollows(userId, page, pageSize)
-                call.respond(
-                    HttpStatusCode.OK,
-                    posts
-                )
-            }
+            val posts = postService.getPostsForFollows(call.userId, page, pageSize)
+            call.respond(
+                HttpStatusCode.OK,
+                posts
+            )
         }
     }
 
@@ -89,35 +71,32 @@ fun Route.getPostsForFollows(
 
 fun Route.deletePost(
     postService: PostService,
-    userService: UserService,
     likeService: LikeService
 ) {
     authenticate {
-        delete ("/api/post/delete") {
+        delete("/api/post/delete") {
             val request = call.receiveOrNull<DeletePostRequest>() ?: kotlin.run {
                 call.respond(HttpStatusCode.BadRequest)
                 return@delete
             }
             val post = postService.getPost(request.postId)
-            if(post == null){
+            if (post == null) {
                 call.respond(
                     HttpStatusCode.NotFound
                 )
                 return@delete
             }
-            ifEmailBelongsToUser(
-                userId = post.userId,
-                validateEmail = userService::doesEmailBelongsToUserId
-            ) {
+            if (post.userId == call.userId) {
                 postService.deletePost(request.postId)
                 likeService.deleteLikesForParent(request.postId)
                 //TODO: Delete comments from post
                 call.respond(HttpStatusCode.OK)
 
+            } else {
+                call.respond(HttpStatusCode.Unauthorized)
             }
 
 
         }
     }
 }
-
